+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * React Native 0.69 support
 * Bumps the minimum supported React Native version to 0.60.0
 * Drops manual linking support
+* Adjusts source maps auto upload script on Android to support the bundled Hermes in RN v0.69
 
 ## 11.0.2 (2022-07-20)